Adds a GitHub action to check proof with Coq on each push
A badge in the README shows the latest build status of the master branch.

The check_proofs action first installs opam and Coq. It then compiles regex.v
with coqc.

The opam root (including the Coq installation) is cached for future runs. The
first run takes about 16mins and subsequent runs using the cache take 30s.

This is a first attempt. An obvious improvement is to use Coq's utility for
autogenerating a Makefile instead of hardcoding the files to check.