Fix Coq 8.13.0 Warnings about hint locality
We can make all of our hints locality export.

The default hint locality outside of a section (where the default is
local) is changing in 8.15 from global to export.

In the meantime hints wihtout a locality setting cause warnings in 8.13
and will be considered an error in 8.14.

“local” means that the hint completely disappears at module closing time
“export” means that the hint is made available whenever its enclosing module is imported
“global” means that the hint is made available whenever its enclosing library (i.e. the toplevel module) is merely required

The global locality causes issues with the modularity of code.

See https://coq.discourse.group/t/change-of-default-locality-for-hint-commands-in-coq-8-13/1140